Add scope based multi tenancy support
A Tenant is identified by some string, currently we're switching
based on request.host but this may change in the future.

When a Tenant is identified the site which it represents is mapped as
current and all queries within the active request are scoped on that
site using default_scope.

Multiple Tenants may represent a single site. This will be useful when
offering custom domains for web facing content but still retaining
customer.xyz.tld for administration activities.

Integrationt to existing models to come.