Enable country records to format addresses
Countries which are activated within the system will supply a postal
address template and method for merging addresses belonging to that
country into a string valid per that countries postal rules.

Population of country details within seeds file needs to occur in the
next commit.

Postal Address formating rules are described indepth here:

http://www.columbia.edu/~kermit/postal.html

In particular refer to the description of the `city line` field which
seems to have the most variance internationally.