Enable domestic and international Adddress formats
Postal carriers generally require labels that are domestic in nature
i.e. Posted within the same country as the destination address, to not
carry a country line after the city line.

Naturally for international destinations the country line is required.

Country address generation now requires the source country to be
specified. For Address this is the associated
Site#distribution_address which is the location that mail/packages
from the Site owner will be sent from. We currently assume only a
single distribution center.