Remove distribution_address concept
On furhter though the last commit handling postage addresses was
starting to drift to far towards managing distribution channels which is not the current intent of this software.

Instead the system now relies only on the subscribers main address
when generating postal addresses and allows a forced override to gain
the international format if required, e.g drop shipping to a local
customer but from an international supplier.