Abstract LineItem and add CommodityLineItem
This change will allow introduction of multiple line item types as
previously highlighted.

Additionallty the naming of `ProductLineItem` was changed for
`CommodityLineItem` so that confusion with existing `Product` and
`Variant` models was not introduced.

In this case we define commodity based on the wikipedia entry as:

    'In economics, a commodity is a marketable item produced to
    satisfy wants or needs. Economic commodities comprise goods and services.'

via http://en.wikipedia.org/wiki/Commodity which fits perfectly as an
abstract term.