Allow VariationInstance to indicate form element
This will allow template authors to write custom logic for `radio` and
`drop_down` methods of presenting content to end users. The shop owner
in turn can select the option that best suits them and the type of
variant data being selected by end users.

A future extension point of the `render_as` enum to include a choice
to display options as clickable images is highly likely.