Don't silently ignore nil assignment to parent
When a mutator accepts a `nil` value and returns without taking any
action, it can mask errors elsewhere in the code. In particular, it can
cover up the piece of code which is incorrectly *retrieving* the `nil`
value to be assigned. If we truly expect never to assign `nil` here, we
should fail hard and fast when/if it happens.

This required a bit of a restructure of the factories, to avoid
`FactoryGirl` needing to assign `nil` to the parent which wasn't being
used. The workaround was to only set each parent in a trait, and don't
pass in a nil value for the missing parent when invoking the factory.