Add ShoppingCart + ShoppingCartOperation models
This models a shopping cart as an append-only queue of transformations
made to the cart. A ShoppingCart object is a single instance of a person
navigating the store to select items for purchase.

Each change to the shopping cart (e.g. add item, update quantity, remove
item, edit metadata) will be represented as a single operation.
Specifics on this to come in a future commit.