Add logic for applying basic cart updates
Each cart operation pertains to a UUID. The UUID refers to a line item
in the cart, so successive changes to the line item will retain the same
UUID. The net effect is essentially that a shopping cart is an
append-only list of operations.

The primary goal here is to retain all operations that were performed on
the cart so we can report on something like an item which was removed
from the cart. The logic attempts not to lose any items that the
customer wanted to add to their cart, so if something funky comes
through from the frontend, the logic will attempt to make it valid and
save it.