Fix "relation does not exist" with Temping tables
Resolves #13

There's no guarantee that the same ActiveRecord connection is used
across different examples, and temporary tables are scoped to the
current session only.

The way Temping works internally, it uses `Object.const_get` to detect
the presence of a temporary model, and `Object.const_set` during
creation. If the constant exists already, the temporary table isn't
created.

The fix is in two parts:

1. Create a new temporary table for each test case, in case the
connection has been reinitialized.
2. Call `Temping.teardown` to clear its internal state, so that it will
not assume the table already exists.