Values for a Variation must be orderable
Previously the set of values associated with a single Variation where
computed from the set of associated VariationInstances but this left
the values in a potentially non natural order when viewed in the front
end e.g.

[xx-large, medium, small, x-large, large]

With this new relationship we can correctly order each value to
product a natural view e.g.

[small, medium, large, x-large, xx-large].