Refine Image, ImageInstance and ImageFile publish
Attempting 'summarise' json structure for the front end is biting us
now that we're starting to move towards a scenario where a summary of
Site data is exported for frontend to use along with specific page
data as we've done until now.

This will only compound when json rendering changes to a referential
implementation in the near future. At that time an ImageInstance for
example will only render the Image id which it uses. Actual Image data
will be supplied within Site data export for reuse.