Refine Category and CategoryMember publish
This change is in the same vein as the previous adjustment within
images.

Attempting to 'summarise' json structure for the front end is biting
us now that we're starting to move towards a scenario where a summary
of Site data is exported for frontend to use along with specific page data as we've done until now.