Refine Product fields inline with usage
Firstly this change migrates Product#description to Product#markdown
through introduction of a new Markup concern and specs.

The existing Product#short_description is then renamed to
Product#description a plain text, preferably short, description of
Product. This is inline with the use of description in categories,
category_members, image_instances and others.