Add tags to ImageInstance
This will allow us to use ImageInstances for a specific
purpose(s) when they are tagged. An example of the thinking here is a
set of ImageInstances attached to some future `Page`. Those that are
tagged `slider` are, unsurprisingly, used by slider macros to generate
the HTML it requires whilst leaving other ImageInstances attached to
the page to be used for other purposes (potentially reused).

More immediently this has a possible application in allowing a Product
with multiple images to highlight an image based on source category
navigation. An example of the thinking here is that someone navigating
to the Product page from category X would see images tagged X or at
least default to an image tagged X where multiple are shown (again in
a slider, what is it with me and slider examples....).