Add Path and start wiring Product to use it
`Path`: Represent alls routable content within
ZW. It makes use of the gem `closure_tree` to structure model data
within a tree format. Being a relative tree noob the following was
most useful when determining choice of tree:
http://www.slideshare.net/billkarwin/models-for-hierarchical-data
along with specifications within the Gem itself.

`Path` is polymorphic so that all routable content, e.g. `Product` and
`Category`, can utilise it when required. The concern `Routable` is
used for all content that has a `Path` to ensure standardisation in this process.

This commit starts to migrate `Product` to use the `Path` model and
`Routable` concern. This will require further refinement before PR.