Make it clear that Preference deals with URI Paths
At the preferences level we are supporting paths which are provided
and stored in URI path format. This should be clear on all counts. Per
PR comments this commit introduces changes to preferences to indicate
that the fields it holds are URI Paths.

Going forward any usage of a uri formatted, string stored path i.e.
`/<segment>/<segment>` will be referenced as xyz_uri_path. All other
path structure data is considered to be in array format, per the
conventions used within closure_tree gem.

A pending commit will flow this change out to the rest of the codebase.