Allow bag to hold multiple commodity instances
This represents a proposed change in functionality for futhter
discussion with @smangelsdorf who knows the internals here better than
I.

Without metadata being considered to uniquely identify a bag
operation, as modified in the previous commit, the following issue was
identitifed.

If a commodity existed in the bag and the user wanted to add
another instance of the commodity, differing only by metadata (in the
specific example a gift card message), the outcome was unexpected. The
result was such that the initial addition metadata was 'overwritten'
with the new metadata, and quantity was added to existing quantity.
Total line items in the bag view remained the same.

With this commit people adding the same commodity multiple times will
see multiple line items. This is less optimal when it is the same
commodity with no metadata differences but is likely what less
experienced users expect to see. It is certainly the expectation if
you're changing metadata on each addition.

Thoughts?