Allow models to describe their extended metadata
This will allow us to store JSON to accurately describe why a
particular model instance has elected to store additional metadata
and details of how it should be presented/collected/managed by other
aspects of the system.

This will be immediently useful for frontend which needs to understand
the extended metadata fields which products want to request from
users. Data such as the type of HTML control, field labels, help etc
will all be stored and supplied to frontend.