Add `use_billing_details_for_shipping` method
When rendering the form for updating the details of an order, Rails'
form helper expects to see an attribute for each form control. Rather
than build the HTML manually, it made more sense to expose this as a
fake attribute so that the initial value could be set correctly in the
form.