Correct error in publishing images spec
During PR a fault was noted within this spec which was not
reproducible. I was also unable to reproduce it.

On investigating the diff of the original error I noted that height
was being set to a word instead of a numeric string.

This corrects that error, but we're still going to need to keep an eye
on this spec in future runs for re-occurance.