Show estimated shipping for bag
When specific configuration and conditions are met it is possible to
show an estimated shipping charge on the bag page.

This may get more advanced in the future if the UI is adapted to
request location details but in the present form will show a shipping
estimate when a shipping rate has been configured as a source that can
be utilised by the bag UI and meets other criteria of the bag at the
current time, namely subtotal + weight.

This is likely to be used by stores that have a standard shipping
option, regardless of destination.