Tweak ShoppingBag
The ShoppingBag workflow has been tweaked in terms of naming. It will
now be considered `active` until the checkout process finalizes it.

The ShoppingBag will also now have a reference to all orders which it
spawns (depending on navigation by the user there may be multiple
orders relating to a single bag, though generally it should be 1:1.)